zzzz

+++ conf/evolutions/default/2.sql
@@ -1,9 +1,9 @@
-# --- New Account Schema
+# --- Account Schema
 
 # --- !Ups
 
 CREATE TABLE account (
-    id         integer NOT NULL PRIMARY KEY,
+    id         bigserial NOT NULL PRIMARY KEY,
     email      varchar NOT NULL UNIQUE,
     password   varchar NOT NULL,
     name       varchar NOT NULL,
@@ -14,7 +14,6 @@ CREATE TABLE account (
 
 DROP TABLE account;
 
-# --- New Account schema
